# backend/contracts/rustrag.openapi.yaml
openapi: 3.1.0
info:
  title: RustRAG API
  version: 0.1.0
  description: |
    Runtime API contract for the current RustRAG backend implementation.

    This contract is hand-maintained from the implemented Axum routes in `backend/src/interfaces/http`.
    It aims to reflect the currently available endpoints, request bodies, response shapes,
    auth requirements, and common error envelopes as accurately as practical.
servers:
  - url: http://localhost:8095
    description: Local default
security:
  - bearerAuth: []
...
paths:
...
  /v1/uploads/ingest:
    post:
      tags: [content]
      summary: Upload a UTF-8 text-like file and ingest it
      operationId: uploadAndIngest
      description: |
        Requires `documents:write`, `workspace:admin`, or an `instance_admin` token.

        Current implementation only supports UTF-8 text-like uploads. Expected multipart fields:
        - `project_id` (required)
        - `source_id` (optional)
        - `title` (optional)
        - `file` (required)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [project_id, file]
              properties:
                project_id:
                  type: string
                  format: uuid
                source_id:
                  type: string
                  format: uuid
                title:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded file ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadIngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/usage-events:
    get:
      tags: [usage]
      summary: List usage events
      operationId: listUsageEvents
      description: Requires `usage:read`, `workspace:admin`, or an `instance_admin` token.
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Usage event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsageEventSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/cost-ledger:
    get:
      tags: [usage]
      summary: List cost ledger entries
      operationId: listCostLedger
      description: Requires `usage:read`, `workspace:admin`, or an `instance_admin` token.
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Cost ledger entry list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostLedgerSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token minted by `POST /v1/auth/tokens`.
        Note: some read/list endpoints are currently public and do not require auth.
  parameters:
...
    projectIdOptional:
      in: query
      name: project_id
      required: false
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
          examples:
            apiErrorEnvelope:
              value:
                error: unauthorized
        text/plain:
          schema:
            type: string
          examples:
            authExtractorError:
              value: missing authorization header
      headers:
        WWW-Authenticate:
          required: false
          schema:
            type: string
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
        text/plain:
          schema:
            type: string
  schemas:
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadIngestResponse:
      type: object
      required: [document_id, external_key, chunk_count]
      properties:
        document_id:
          type: string
          format: uuid
        external_key:
          type: string
        chunk_count:
          type: integer
          minimum: 0
        mime_type:
          type: [string, 'null']
...
    UsageEventSummary:
      type: object
      required: [id, usage_kind, raw_usage_json]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: [string, 'null']
          format: uuid
        model_profile_id:
          type: [string, 'null']
          format: uuid
        usage_kind:
          type: string
          example: query
        prompt_tokens:
          type: [integer, 'null']
          format: int32
        completion_tokens:
          type: [integer, 'null']
          format: int32
        total_tokens:
          type: [integer, 'null']
          format: int32
        raw_usage_json:
          type: object
          additionalProperties: true
    CostLedgerSummary:
      type: object
      required: [id, usage_event_id, provider_kind, model_name, currency, estimated_cost, pricing_snapshot_json]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: [string, 'null']
          format: uuid
        usage_event_id:
          type: string
          format: uuid
        provider_kind:
          type: string
        model_name:
          type: string
        currency:
          type: string
          example: USD
        estimated_cost:
          type: number
          format: double
        pricing_snapshot_json:
          type: object
          additionalProperties: true